ConnectionManager::shutdown should not throw any exception and ClusterTest.testListenersWhenClusterDown fails (#297)
* Prevent ConnectionManager::shutdown to throw any exception.

Reverted the order of InvocationService shutdown and ConnectionManager shutdown in order to eliminate any unnecessary retries happening during client shutdown, and prevent these retries causing any exceptions. Note that InvocationService::shutdown is only setting a boolean flag.

At Java client, there is a CleanResourcesTask which periodically performs the cleanup of invocations. We really need such a task implemented in the future.

* Corrected the return value of InvocationService::removeEventHandler method.

Fix: Set the handler registration id correctly, previously we left it as unset (empty) which indirectly caused the removals of the registration to fail.

* Removed the unneeded isOpen flag from InvocationService, it uses LifecycleService::isOpen instead. Also, suppressed some logs printed during shutdown. Updated the start sequence to start the InvocationService start before ConnectionManager start (currently InvocationService::start is no-op). Updated shutdown sequence to shutdown PartitionService before ConnectionManager.