Changes per review comment.
Removed the create method from the SerializerBase. Instead introduced the StreamSerializer which has the read method similar to Java implementation. If the user want to add custom serializer factory it needs to provide a class that implements the StreamSerializer.

I had to also change the DataSerializer and PortableSerializer so that their read method receives the object instance to be used in case that no factory was provided.