Client may crash during shutdown when ShutdownTask is being run (#430)
Do not access any client object after ShutdownTask runnable is finished. Obtain the client name at the constructor of ShutdownTask to eliminate the illegal memory access when the client is shutdown e.g. during heartbeat failure and the client object is destructed already (the destructor waits on a latch which is released once the ShutdownTask::run is finished. Hence the ShutdownTask thread should not access the possibly destructed client on thread destructor where the name of the thread is written.

fixes #428
fixes #429

Cleaned up unused header file and corrected the exception print at the ClientPartitionServiceImpl.