Corrected the _ELPP_DEBUG_LOG macro setting when the build type is "Release". This flag was being defined to 0 when Release builds are compiled which caused the LoggerConfigFromFileTest.testFinest to fail. By defining this flag to 1 even when "Release" builds are compiled, the test passes now.
Modified LoggerConfigFromFileTest so that the log output file is truncated (contents cleared) at every test case start.

Changed the ClientTestSupport logger from protected to private and only let derived test classes use getLogger() API instead.

Also increased the ROLL_OUT_SIZE parameter for the LoggerConfigFromFileTest to 2MB from 17 bytes. The roll-out does not work as expected in environments, and hence i am not testing for roll-out, this is a bug of the library, it usually rolls out when size is a lot greater than 17 bytes.