Changed the ClientMapTest to be parametrized rather than type parametrized which simplifies the test code.
ClientConnectionManagerImpl shutdown is changed to shutdown the pendingConnections as well and closes the threads and shutdown the connection strategy before closing connections so that a connection close will not retry to open a new connection during shutdown. This causes deadlocks.

The pending AuthenticationFutures are also set with exception that client is shutting down. The SimpleExecutorService also now processes any outstanding runnable in the queue during shutdown.