Removed the header-only asio and switched to boost::asio since the boost::asio no longer requires boost::system library compilation and linkage but it can now be used as header-only (See https://www.boost.org/doc/libs/1_66_0/libs/system/doc/reference.html https://stackoverflow.com/questions/31110409/boost-error-code-header-only-does-not-have-the-advertised-effect/31131980#31131980).
Also started to use find_package for boost (This way it will be more flexible and compatible if the user wants to compile and use with their own version of boost). It looks at the component installations for asio and property_tree first and if it can not find them as components, it searches for a complete boost installation.