Removed submodules.
Boost is being searched first in user's cmake environment, if it is not found, it is being downloaded and unzipped and the unzipped version is being used.

Googletest is also being searched first and if it is not installed, then it is added as a cmake external project so that it is also being downloaded buring configuration of cmake.