Architectural changes to use boost thread_pool, boost::future, std::promise, std::nested_exception, connection local invocation and event handler maps (#584)
* Changed the event processing executor usage to use asio::thread_pool.

Changed the ResponseThread to ResponseProcessor and it now processes the response message directly by default and will use a thread_pool if the newly introduced parameter `hazelcast.client.response.executor.thread.count` is set to a positive value.

Removed custom Thread, Future implementations and used std::thread, boost::asio::thread_pool and boost::future/boost::promise.

boost::future supports continuation (`.then` support).

Changed the invocation and event handler map design. Using an `unordered_map` which which only being accessed from the same `io` thread for the connection. This eliminated the need for using sync maps for invocations and events. If any other thread needs to access the invocations, it passes this request to io thread. This is only needed when closing the socket for notifying the outstanding invocations. If an invocation can not be queued into the connection (i.e. no connection exists for the invocation and it is being retried), then it is registered in a sync map at the invocation service only for the first retry and cleared from the service when a response is finally received. This ensures that the user thread is always being notified.

Exception architecture is compeletely changed so that we use std exception_ptr capabilities. Also integrated std::nested exceptions for exceptions which has a cause exception. The cause is obtained by using `std::rethrow_if_nested` method.

The project now depends on boost::thread for boost::future. Changed cmake so that the boost installation is user's responsibility. Also using the boost Thread library version latest which is version 5.

* Linux build fixes.

* Fixed the examples per changes in the APIs.

* linux test fix.

* Removed custom implementations of `CountDownLatch, Mutex, LockGuard, ConditionVariable`. Replaced them with standard and boost based implementations.

Changed required boost version to 1.72.

* Fix to `IExecutorService`.

* Fix to code examples.

Removed unneeded include for windows (`WinSock2.h`) which causes error in compilation.

* Test fix (count down latch count was incorrect).

Windows Windsock2.h usage for easylogging to work with boost.

* test fixes.

* Eliminated the extra invocations registration at the invocation service.

* Removed all invocation and even handler registrations in listener and invocation service.

Changed the CallIdSequence to use std::array of atomic integers.

* Put `boost::` namespace to all implicit boost usages which was needed for compiling successfully at Windows (fails due to `ambiguous symbol` error which can be both from std or boost namespaces).

* Fixed the `boost::latch` usages in the examples.

* test fix for wait time.

* Removed unneeded AtomicArray class.

* fix for issue https://github.com/chriskohlhoff/asio/issues/431

Corrected the time string print utility.

Replaced some of the time usages with std::chrono related ones.

Made the timer cancel operations non-exception throwing.

Replaced the system_clock usages with steady_clock usages.

Warning fixes for windows for unused exceptions.

* Added `hz_thread_pool` for graceful shutdown of thread pools.

Some fixes so that the exceptions do not abort the io threads. Also, destroying the timer and pools upon shutdown so that no associated resource is left (which may happen for windows).