All proxy getters are made async by returning boost::future.
A minor fix to `DefaultNearCache` where the name was not cached correctly and an invalid reference was being kept, which caused the address sanitizer fail during the tests on llvm (mac OS).

Windows build failure fix (dll symbol error). Carried the implementsation of reliable_listener constructor from header to cpp.