Removed destruction of the thread_pool on shutdown and left it to actual client destruction. Failing to do this causes illegal memory access during destruction of stands in the listener_service_impl.
Removed the work around of bug https://github.com/chriskohlhoff/asio/issues/431.

Added correct handling of the invocation retry timer. If it is not assigned to a timer variable in the invocation then it will actually be destroyed immediately, which will cause the timer cancelled immediately on destruction but this is not what we want.