Make lib and cmake target names the same as repository name, do NOT support cmake install multiple configurations simultaneously (#812)
* It is a problem for users when out library and github name is different from the cmake find_package name and library names. It becomes too confusing to the users and it is NOT recommended by package managers such as conan and vcpkg. Also, supporting different targets for ssl and static was not a good idea and source of confusion. The user configures it anyway when using cmake or package managers, hence, deleted those options.

Eliminated the generation of different targets for static and ssl and it will be only controlled by cmake options. Only a single type is supported per installation. The library and target names are also changed to the same name as the project `hazelcast-cpp-client`.

* Changed cmake option `BUILD_SHARED_LIB` to `BUILD_SHARED_LIBS` as this is the standard recommended option name https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html

Removed all the `BUILD_STATIC_LIB` flags and related documentation.

Updated the scripts to work with the latest changes.

Minor fixes for the copy elision warning of std::move statements.

* Corrected the produced cmake config destination directories.

* Added path `%BUILD_DIR%\bin\%BUILD_CONFIGURATION%` when running the windows test so that gtest.dll can be found during running the test executable.