Fix PR Builder failing to start (#1231)
The PR builder [is failing to even start the
tests](https://github.com/hazelcast/hazelcast-cpp-client/issues/1222)
due to remote controller now requiring Java 17.

This is non-trivial to fix, as several other things have _also_ broken
in the meantime.

### Force-merge required:
`.github/workflows/build-pr.yml` is configured to use use
`pull_request_target`, rather than `pull_request` - to allow execution
of the action _on the target repository/branch_, allowing it's secrets
to be accessed.
The problem with this is what when these scripts contain an issue,
**even if modified in a PR, the _existing_ scripts will still be
executed** - meaning it's **impossible** to get the PR builder to pass,
whatever you modify.
As such, I've executed this PR in a branch (over a fork) and amended
this behavior, so that the execution of this PR build can be properly
observed in a [test run on a previous
revision](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/10506067493).
This PR will then require a force-merge.
Ideally this behaviour should change, but out-of-scope for this PR.

### Changes:
- Centralised configuration _where possible_
   - Java version
   - Java distribution
   - Hazelcast version
- Upgraded Java runtimes to Java 17
- Upgraded Hazelcast dependency from `5.3.0-SNAPSHOT` to `5.3.0`
   - the `SNAPSHOT` is no longer available
- this is the nearest (although now outdated) direct replacement (can
and should be upgraded -
[separately](https://github.com/hazelcast/hazelcast-cpp-client/issues/1225))
- Explicitly install the right Java / Maven versions on the runners, in
favour of the bundled (outdated) versions
- Refactor use of `maven-dependency-plugin` to specify a version
- Currently implicitly uses the version bundled with Maven - but this
has broken following Maven upgrade
- Picks a version (not the latest!) that supports a custom download
destination and a custom repository _simultaneously_
- Fix permissions issue installing `boost` on OSX
- Fix `error: unqualified call to 'std::move'` by qualifying call

Fixes: https://github.com/hazelcast/hazelcast-cpp-client/issues/1222,
https://github.com/hazelcast/hazelcast-cpp-client/issues/1229,
https://github.com/hazelcast/hazelcast-cpp-client/pull/1207,
https://github.com/hazelcast/hazelcast-cpp-client/pull/1226

---------

Co-authored-by: ihsan <ihsandemir@gmail.com>