Use consistent property for Hazelcast version specification (#1275)
When building, we specify the Hazelcast (platform) version via
`config.env`:
https://github.com/hazelcast/hazelcast-cpp-client/blob/a48374e064e87febb3feb274a53170175b80bfdd/.github/config.env#L4

Specifically, in a `HAZELCAST_VERSION` property.

In `start-rc` we do something similar, with a default value if not
provided for local execution:
https://github.com/hazelcast/hazelcast-cpp-client/blob/a48374e064e87febb3feb274a53170175b80bfdd/scripts/start-rc.sh#L37

Note the property there is `HZ_VERSION`.

This means if you try to specify a _different_ Hazelcast version by
changing the config, the `start-rc` script fails as it can't find the
`hazelcast-enterprise-???-tests.jar` file.

Updated to rename the `HAZELCAST_VERSION` property as it [was introduced
more
recently](https://github.com/hazelcast/hazelcast-cpp-client/pull/1231).