Simplify `Debug`/`Release` options (#1280)
The change in
https://github.com/hazelcast/hazelcast-cpp-client/pull/1279 made it
obvious that adding configuration for `Debug` builds (e.g. "add an extra
compilation argument") was very tedious as it needs to be declared in
_each_ workflow and then passed all the way down to the actual scripts
that build the code.

Instead, it'd be much cleaner if the jobs _only_ declare a release type,
and the build scripts react to that with the appropriate actions (e.g.
adding some extra debug or whatever).

Changes:
- migrate the _implementation_ of the `DEBUG` builds into the `build-`
scripts
- e.g. `WARN_AS_ERROR` can be assumed when `DEBUG` is enabled, rather
than being duplicated in each job configuration
   - remove the architecture to pass these redundant params down
- pass `BUILD_TYPE` into the `build-` scripts as an environment
variable, as already used for other parameters and for the Windows
equivalent (`BUILD_CONFIGURATION`)

This PR is the refactor-half of
https://github.com/hazelcast/hazelcast-cpp-client/pull/1279, split out
separately to make that PR smaller.