Fixed build problems with latest Boost library versions.
Updated the project to use the minimum C++ standard to be the same as the one required by the Boost library. This means if a user is using an older Boost version we use C++11 but if newer Boost version is used we change level to C++ 14 or C++ 17. the algorithm is like this:
```
    # Check Boost version and set C++ standard accordingly
    # Our library can not use a lower C++ standard than the one required by Boost.
    if (Boost_VERSION VERSION_LESS 1.70)
        set(CMAKE_CXX_STANDARD 11)
    elseif (Boost_VERSION VERSION_LESS 1.77)
        set(CMAKE_CXX_STANDARD 14)
    else()
        set(CMAKE_CXX_STANDARD 17)  # Boost.JSON, Boost.Process, etc.
    endif()
```