Fix inconsistent usage of environment variable character case (capital/small case) in workflow actions (#1379)
Inconsistent usage of environment variable character case (capital/small
case) caused the tests not to run.

Changed all the environment variables to be upper case. The tests should
now work as expected in the builed including PR builder.

Triggered a sample run
[here](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20297443182)
for PR #1376 to test the changes. This run does not include the change
to use RUN_TESTS=true in pr builder, hence, we should see that it still
should not run the tests.

In the second commit of the PR I enabled the RUN_TESTS=true in pr
builder and triggered another workflow run
[here](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20297609221)
where I run this PR #1379 and we should see the tests are run. As you
can see, the tests are actually run e.g.
[here](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20297609221/job/58295008466)

I had to force merge since it involves PR builder change.