Codecov publish fix [HZ-5302] (#1381)
Codecov publish fix. Avoiding to do this caused the publish action find
and use extra coverage files for the test files which were excluded when
running the gcov.

If you look at the [current
run](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20331270807/job/58406882376)
and download the
[coverage_report.zip](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20331270807/artifacts/4909506075)
you can see that the test directory is actually excluded but it is
published differently to codecov.

The details on the config option `disable_search` is
[here](https://github.com/codecov/codecov-action/blob/671740ac38dd9b0130fbe1cec585b89eea48d3de/action.yml#L26)
more details
[here](https://github.com/codecov/codecov-action/blob/671740ac38dd9b0130fbe1cec585b89eea48d3de/action.yml#L26).

fixes #1308