Enforce code formatting in PR via `git-clang-client` (#1383)
In https://github.com/hazelcast/hazelcast-cpp-client/pull/946 the
codebase was reformatted for consistency, but as this was not enforced
the codebase has drifted since then.

To avoid this happening again, enforce this at the PR stage.

Testing:
- [example
failure](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20455944532/job/58777930370)
- [example success

](https://github.com/hazelcast/hazelcast-cpp-client/actions/runs/20455961135/job/58777982668)

This is alternative implementation addressing feedback from
https://github.com/hazelcast/hazelcast-cpp-client/pull/1342#issuecomment-3580258782

Closes https://github.com/hazelcast/hazelcast-cpp-client/pull/1342

---------

Co-authored-by: ihsan demir <ihsandemir@gmail.com>