Removed individual warning suppressions in each file for C4251 [HZ-5297] (#1387)
Removed individual warning suppressions in each file for C4251 dll
export warning in wondows and put a global suppression in Cmake file to
suppress it for the project wide.

It is OK to disable this warning project wise, because it means warnings
about unexported fields such as std::vector in a class but it is not
important because our users will compile from source code and hence they
will always use the same e.g. `std::vector` versions.

This makes less code lines for the project and remove the unneeded
repetitions.