Fixed ClientMapTest.testPutIfAbsentTtl so that it is in sync with Java test https://github.com/hazelcast/hazelcast-mono/blob/aad45696a4bb784c6a538e22d26f183973ace9c4/hazelcast/hazelcast/src/test/java/com/hazelcast/client/map/ClientMapTest.java#L542.
The problem was that we dependent on timing on the entry expiry and 1 second may not be enough in busy times. Hence, it is changed to 10 seconds. We also re-put the entry with 10 seconds ttl in the send put also to avoid expiry when we checked the value in the next step.

fixes #1145